Refactor random number generation for planet modeling to improve FP32 behavior
Updated the PlanetModelingManager to implement a tie-break mechanism in the probability checks, ensuring consistent behavior at boundary values. Changed random number generation in various classes to utilize double precision for critical calculations, enhancing accuracy and reducing edge case issues. Adjusted related classes to maintain compatibility with the new logic, including updates to CudaPlanetNative and UniverseGenF32 for improved performance in planet generation.

+++ SeedCli/CudaPlanetNative.cs
@@ -52,10 +52,10 @@ namespace SeedCli
             public float temperature_bias;
             public float radius;
             public float luminosity;
-            public float rand1;
-            public float rand2;
-            public float rand3;
-            public float rand4;
+            public double rand1;
+            public double rand2;
+            public double rand3;
+            public double rand4;
             public int theme_seed;
             public int type_case;
             public int singularity_flags;
